added types for weapons, eggs, pets, mounts and hatching potions
