+++ src/types/manifest.ts
@@ -1,3 +1,5 @@
+import { EggItemMeta, HatchingPotionItemMeta, ItemMeta, StableItemMeta, WeaponItemMeta } from "./item-meta";
+
 export type ImageMeta = {
   fileName: string;
   width: number;
@@ -5,20 +7,9 @@ export type ImageMeta = {
   format: 'png' | 'gif';
 };
 
-export type ItemMeta = {
-  key: string;
-  text: string;
-  imageFileNames: string[]; // reference to ImageMeta.fileName
-  notes?: string;
-  price?: number;
-  currency?: string;
-  set?: string;
-  twoHanded?: boolean;
-};
-
 export type GearItems = {
   sets: Record<string, string[]>; // set name as key
-  weapon: Record<string, ItemMeta>;
+  weapon: Record<string, WeaponItemMeta>;
   armor: Record<string, ItemMeta>;
   head: Record<string, ItemMeta>;
   shield: Record<string, ItemMeta>;
@@ -54,11 +45,11 @@ export type PetTree = {
 export type AvatarManifestItems = {
   background: Record<string, ItemMeta>;
   gear: GearItems;
-  egg: Record<string, ImageMeta>;
-  hatchingPotion: Record<string, ItemMeta>;
-  pet: Record<string, ItemMeta>;
+  egg: Record<string, EggItemMeta>;
+  hatchingPotion: Record<string, HatchingPotionItemMeta>;
+  pet: Record<string, StableItemMeta>;
   petTree: PetTree;
-  mount: Record<string, ItemMeta>;
+  mount: Record<string, StableItemMeta>;
   hair: HairItems;
   skin: Record<string, ItemMeta>;
   body: BodyItems;
